monitorx_backend/requirements.txt
amqp==5.1.1
arrow==1.2.3
asgiref==3.5.2
autopep8==1.7.0
billiard==3.6.4.0
cachetools==5.3.0
celery==5.1.0
certifi==2021.5.30
cffi==1.15.1
chardet==4.0.0
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.4
Django==3.2.4
django-celery-results==2.1.0
django-environ==0.4.5
django-filter==2.4.0
django-guardian==2.4.0
django-prometheus==2.1.0
django-redis==5.0.0
djangorestframework==3.12.4
google-auth==2.16.0
idna==2.10
itsdangerous==2.0.1
jmespath==0.10.0
kombu==5.2.4
kubernetes==25.3.0
multi-key-dict==2.0.3
numpy==1.23.3
oauthlib==3.2.2
pandas==1.5.0
pbr==5.11.1
prometheus-client==0.11.0
prompt-toolkit==3.0.18
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-gitlab==3.13.0
python-jenkins==1.7.0
pytz==2021.1
PyYAML==6.0.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
tld==0.12.6
toml==0.10.2
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.2.1
